requirements.txt
asgiref==3.6.0
beautifulsoup4==4.12.2
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.2
django-jazzmin==2.6.0
django-simple-history==3.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
environs==9.5.0
et-xmlfile==1.1.0
excel2img==1.4.0
idna==3.4
inflection==0.5.1
marshmallow==3.19.0
openpyxl==3.1.2
packaging==23.1
Pillow==9.5.0
PyJWT==2.6.0
pypiwin32==223
python-dotenv==1.0.0
pytz==2023.3
pywin32==306
PyYAML==6.0
requests==2.29.0
soupsieve==2.4.1
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
